First steps on porting over Untyped.Pure
The idea is instead of having separate modules with similarly named
types, we have a data family such that e.g. 'Parsed Struct' is the
"pure" version of struct -- we'll use this for typed types too.

So far this only defines an instance for capabilities; we'll work our
way up the dependency graph from the leaves.